README
To use the application, start node server on port 3000 and navigate to localhost:3000. The application will present the index page or welcome page. Here it displays the last five messages that were sent in the messaging board. I made up the name : connectimus for the application. A link on the index page ‘ChatRoom’ will take the user to the Chat room to post messages.
I have used two routes namely blog and index for this application. Also I have used two jade templates: blog/basic.jade for the chatroom and index.jade for the home page. My client side javascript is also contained within a namespace: customVars. For the 300 character validation, I used jquery onkeydown event to trigger a validation everytime a button is pressed. The validation process involves fetching the current length of the message text box and checking if the length is more than 300. If it is more than 300, a message is displayed and the post message button is disabled. If the characters are less, the post message button remains active. Normal validations that check for empty text boxes also ensure that the user enters his name and message before posting. Once the user successfully posts a message, it is stored in the database along with the time stamp, and displayed in the chat area in real time too.